Add common interface for snapshot and restore request
